requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.0
kafka-python==2.0.2
psycopg2-binary==2.9.7
pydantic==2.3.0
SQLAlchemy==2.0.20
pyarrow==13.0.0
python-dotenv==1.0.0